leaner head; dynamically load correct language for dropdown menu
