+++ index.html
@@ -5,28 +5,8 @@
 
 <head>
   <title>prosodylab</title>
-
   <meta charset="utf-8">
-
-  <script src="prosodylab/loadScripts.js"></script>
-
-  <script src="javascripts/jquery.min.js"></script>
-  <script src="javascripts/jquery-ui.min.js"></script>
-
-  <!-- country/region select -->
-
-  <link rel="stylesheet" href="javascripts/jquery-ui.css">
-  <link rel="stylesheet"
-    href="javascripts/node_modules/country-region-dropdown-menu/assets/css/geodatasource-countryflag.css" />
-  <link rel="gettext" type="application/x-po"
-    href="javascripts/node_modules/country-region-dropdown-menu/languages/en/LC_MESSAGES/en.po" />
-  <!-- Japanese country options -->
-  <link rel="gettext" type="application/x-po"
-    href="javascripts/node_modules/country-region-dropdown-menu/languages/ja/LC_MESSAGES/ja.po" />
-
-  <!-- css -->
-  <link href="javascripts/jspsych-6.1.0/css/jspsych.css" rel="stylesheet" type="text/css">
-  </link>
+  <script src="prosodylab/setupHead.js"></script>
 </head>
 
 <body></body>
@@ -81,13 +61,13 @@
     fullScreen: false, // doesn't work with IE, but will simply proceed anyway
     //
     // Switch certain components of study on or off
-    hello: true,
-    consent: true,
+    hello: false,
+    consent: false,
     languageQuestionnaire: true,
-    soundCheck: true,
-    micCheck: true,
+    soundCheck: false,
+    micCheck: false,
     headphoneScreener: {
-      includeHeadPhoneScreener: true,
+      includeHeadPhoneScreener: false,
       stimuli: 'original', // options: ['short','original'],
       numberChances: 2, // how many attempts do participants get?
       threshold: 5 / 6, // if equal or higher, participant passes,
@@ -107,10 +87,10 @@
     //
   };
 
+  // load messages for country and region dropdown in correct language
+  setup.loadGeoDataMessages('en');
 
-  // Load scripts and create experiment
-
-  Promise.all(setup.promises(study.language))
+  Promise.all(setup.scriptPromises(study.language))
     .then(function () {
 
       // load messages for interaction with participant:
@@ -373,6 +353,8 @@
      .catch(function (script) {
        console.log(script + ' failed to load');
      }); */
+
+
 
 </script>
