+++ prosodylab/languageQuestionnaire/languageQuestionnaire_en.html
@@ -15,9 +15,9 @@
 
 <p><div>
         <b>Country you grew up in: &nbsp;</b>
-        <select class="gds-cr" country-data-region-id="countryGrowingup" name= "countryGrowingup data-language="en" required ></select>&nbsp; &nbsp;
+        <select class="gds-cr" country-data-region-id="countryGrowingup" name= "countryGrowingup" data-language="en" required ></select>&nbsp; &nbsp;
         <b> Region: &nbsp;</b>
-        <select id="countryGrowingup" name="regionGrowingup" required "></select>
+        <select id="countryGrowingup" name="regionGrowingup" required></select>
     </div>
 </p>
